refactor: 서비스 구조 변경
